# web/src/swagger/swagger-docs.yaml
openapi: 3.0.1
info:
  title: Trisul Service
  description: Trisul Service
  version: "0.1"
servers:
  - url: http://localhost:2041/trisul
    description: Generated server url
...
paths:
...
  /service/static-data/tiles:
    get:
      tags:
        - Static Data Service
      summary: Tile
      description: It will give you list of configured Tile.
      operationId: getConfiguredTile
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Operation Successful
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagedContent'
  /service/static-data/states/{countryId}:
    get:
      tags:
        - Static Data Service
      summary: States
      description: It will give you list of States.
      operationId: getStatesByCountryId
      parameters:
        - name: countryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Operation Successful
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CscDetail'
  /service/static-data/country/{countryKey}:
    get:
      tags:
        - Static Data Service
      summary: Country
      description: It will give you country detail by country code key.
      operationId: getCountryByCountryKey
      parameters:
        - name: countryKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Operation Successful
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CscDetail'
  /service/static-data/countries:
    get:
      tags:
        - Static Data Service
      summary: Countries
      description: It will give you list of countries.
      operationId: getCountries
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Operation Successful
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CscDetail'
  /service/static-data/codes/{codeType}:
    get:
      tags:
        - Static Data Service
      summary: Code
      description: It will give you list of code by code type.
      operationId: getCodeType
      parameters:
        - name: codeType
          in: path
          required: true
          schema:
            type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Operation Successful
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CodeDetail'
  /service/static-data/code/{codeKey}:
    get:
      tags:
        - Static Data Service
      summary: Code
      description: It will give you code detail by code key.
      operationId: getCode
      parameters:
        - name: codeKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Operation Successful
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CodeDetail'
  /service/static-data/cities/{stateId}:
    get:
      tags:
        - Static Data Service
      summary: Cities
      description: It will give you list of Cities.
      operationId: getCitiesByStateId
      parameters:
        - name: stateId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Operation Successful
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CscDetail'
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        description:
          type: string
    ErrorMessage:
      type: object
      properties:
        status:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        error:
          $ref: '#/components/schemas/Error'
...
    CodeDetail:
      required:
        - codeDescription
        - codeID
        - codeKey
        - codePriority
        - codeShortDescription
        - codeType
      type: object
      properties:
        codeID:
          type: integer
          description: Provide code id
          format: int64
        codeKey:
          type: string
          description: Provide code key
          format: string
        codeType:
          type: string
          description: Provide code type
          format: string
        codeShortDescription:
          type: string
          description: Provide code short description
          format: string
        codeDescription:
          type: string
          description: Provide code description
          format: string
        codePriority:
          type: integer
          description: Provide code priority
          format: int64
      description: Code Detail
    CscDetail:
      required:
        - cscCodeID
        - cscDescription
        - cscID
        - cscKey
        - cscPriority
        - cscShortDescription
        - cscType
      type: object
      properties:
        cscID:
          type: integer
          description: Provide id
          format: int64
        cscKey:
          type: string
          description: Provide key
          format: string
        cscShortDescription:
          type: string
          description: Provide short description
          format: string
        cscDescription:
          type: string
          description: Provide description
          format: string
        cscPriority:
          type: integer
          description: Provide priority
          format: int64
        cscCodeID:
          type: integer
          description: Provide code id
          format: int64
        cscType:
          type: string
          description: Provide type
          enum:
            - COUNTRY
            - STATE
            - CITY
      description: Csc Detail
...
    ManagedContent:
      type: object
      properties:
        html:
          type: string
          description: html
          format: string
        tile:
          $ref: '#/components/schemas/ManagedContentTile'
      description: Managed Content
    ManagedContentTile:
      type: object
      properties:
        key:
          type: string
          description: key
          format: string
        type:
          type: string
          description: type
          format: string
        title:
          type: string
          description: title
          format: string
        description:
          type: string
          description: description
          format: string
        icon:
          type: string
          description: icon
          format: string
        isDelete:
          type: boolean
          description: isDelete
          format: boolean
        isAccessed:
          type: boolean
          description: isAccessed
          format: boolean
      description: Managed Content Tile
  securitySchemes:
    Authorization:
      type: http
      description: "**The API operations may require HTTP Bearer authentication scheme\
        \ with JWT format as defined in Open API 3.0.**\nAn Authorization header must\
        \ be present if the operation requires authorization, \n\nfor example:\n```\n\
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c\n\
        ```\n"
      scheme: bearer
      bearerFormat: JWT